refactor configure tests to collect tests and content packs
