+++ Tests/scripts/collect_tests_and_content_packs.py
