+++ Tests/scripts/infrastructure_tests/test_collect_tests_and_content_packs.py
@@ -2,9 +2,11 @@ import json
 import copy
 import os
 from ruamel.yaml import YAML
-from Tests.scripts.configure_tests import get_test_list_and_content_packs_to_install, get_modified_files, RANDOM_TESTS_NUM, TestConf, \
-    create_filter_envs_file
 import demisto_sdk.commands.common.tools as demisto_sdk_tools
+
+from Tests.scripts.collect_tests_and_content_packs import (
+    RANDOM_TESTS_NUM, get_modified_files,
+    get_test_list_and_content_packs_to_install)
 
 with open('Tests/scripts/infrastructure_tests/tests_data/mock_id_set.json', 'r') as mock_id_set_f:
     MOCK_ID_SET = json.load(mock_id_set_f)
@@ -473,8 +475,10 @@ TWO_BEFORE_GA_VERSION = '4.5.0'
 def get_mock_test_list(two_before_ga=TWO_BEFORE_GA_VERSION, get_modified_files_ret=None, mocker=None, git_diff_ret=''):
     branch_name = 'BranchA'
     if get_modified_files_ret is not None:
-        mocker.patch('Tests.scripts.configure_tests.get_modified_files', return_value=get_modified_files_ret)
-
+        mocker.patch(
+            'Tests.scripts.collect_tests_and_content_packs.get_modified_files',
+            return_value=get_modified_files_ret
+        )
     tests = get_test_list_and_content_packs_to_install(
         git_diff_ret, branch_name, two_before_ga, id_set=MOCK_ID_SET, conf=TestConf(MOCK_CONF
     ))
